+++ R/read_log.R
@@ -9,8 +9,9 @@
 #' @examples
 #' read_log(readr_example("example.log"))
 read_log <- function(file, col_names = FALSE, col_types = NULL,
+                     trim_ws = TRUE,
                      skip = 0, n_max = Inf, progress = show_progress()) {
-  tokenizer <- tokenizer_log()
+  tokenizer <- tokenizer_log(trim_ws = trim_ws)
   read_delimited(file, tokenizer,
     col_names = col_names, col_types = col_types,
     skip = skip, n_max = n_max, progress = progress